+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/edit/details/main.js
@@ -15,10 +15,7 @@ define([
 
     'use strict';
 
-    var defaults = {
-            headline: 'contact.accounts.title'
-        },
-        fields = ['urls', 'emails', 'faxes', 'phones', 'notes', 'addresses'],
+    var fields = ['urls', 'emails', 'faxes', 'phones', 'notes', 'addresses'],
 
         constants = {
             tagsId: '#tags',
@@ -26,7 +23,9 @@ define([
             bankAccountAddId: '#bank-account-add',
             addAddressWrapper: '.grid-row',
             addBankAccountsWrapper: '.grid-row',
-            editFormSelector: '#contact-edit-form'
+            editFormSelector: '#contact-edit-form',
+            formSelector: '#contact-form',
+            formContactFields: '#contact-fields'
         },
 
         customTemplates = {
@@ -47,8 +46,11 @@ define([
         };
 
     return {
-
         view: true,
+
+        tabOptions: {
+            noTitle: true
+        },
 
         layout: function() {
             return {
@@ -68,23 +70,18 @@ define([
         templates: ['/admin/contact/template/account/form'],
 
         initialize: function() {
-            this.options = this.sandbox.util.extend(true, {}, defaults, this.options);
-            
-            this.form = '#contact-form';
-            this.formContactFields = '#contact-fields';
+            this.data = this.options.data();
+
             this.autoCompleteInstanceName = 'contacts-';
             this.dfdListenForChange = this.sandbox.data.deferred();
             this.dfdFormIsSet = this.sandbox.data.deferred();
 
-            AccountManager.loadOrNew(this.options.id).then(function(data) {
-                this.data = data;
-                this.render();
-                this.listenForChange();
+            this.render();
+            this.listenForChange();
 
-                if (!!this.data && !!this.data.id && WidgetGroups.exists('account-detail')) {
-                    this.initSidebar('/admin/widget-groups/account-detail?account=', this.data.id);
-                }
-            }.bind(this));
+            if (!!this.data && !!this.data.id && WidgetGroups.exists('account-detail')) {
+                this.initSidebar('/admin/widget-groups/account-detail?account=', this.data.id);
+            }
         },
 
         destroy: function() {
@@ -97,49 +94,18 @@ define([
         },
 
         render: function() {
-            var data, excludeItem, options;
-
             this.sandbox.emit(this.options.disablerToggler + '.change', !!this.data.disabled);
             this.sandbox.emit('sulu.header.toolbar.item.show', 'disabler');
-
             this.sandbox.once('sulu.contacts.set-defaults', this.setDefaults.bind(this));
             this.sandbox.once('sulu.contacts.set-types', this.setTypes.bind(this));
 
             this.html(this.renderTemplate('/admin/contact/template/account/form'));
 
-            this.titleField = this.$find('#name');
-
-            data = this.initContactData();
-
-            excludeItem = [];
-            if (!!this.data.id) {
-                excludeItem.push({id: this.data.id});
-            }
-
-            options = Config.get('sulucontact.components.autocomplete.default.account');
-            options.el = '#company';
-            options.value = !!data.parent ? data.parent : null;
-            options.instanceName = 'companyAccount' + data.id;
-
-            this.sandbox.start([
-                {
-                    name: 'auto-complete@husky',
-                    options: options
-                },
-                {
-                    name: 'input@husky',
-                    options: {
-                        el: '#vat',
-                        instanceName: 'vat-input',
-                        value: !!data.uid ? data.uid : ''
-                    }
-                }
-            ]);
-
-            this.initForm(data);
+            var formData = this.initContactData();
+            this.initForm(formData);
             this.setTags();
             this.bindCustomEvents();
-            this.bindTagEvents(data);
+            this.bindTagEvents(formData);
         },
 
         /**
@@ -275,13 +241,33 @@ define([
         },
 
         initForm: function(data) {
+            var options = Config.get('sulucontact.components.autocomplete.default.account');
+            options.el = '#company';
+            options.value = !!data.parent ? data.parent : null;
+            options.instanceName = 'companyAccount' + data.id;
+
+            this.sandbox.start([
+                {
+                    name: 'auto-complete@husky',
+                    options: options
+                },
+                {
+                    name: 'input@husky',
+                    options: {
+                        el: '#vat',
+                        instanceName: 'vat-input',
+                        value: !!data.uid ? data.uid : ''
+                    }
+                }
+            ]);
+
             this.numberOfAddresses = data.addresses.length;
             this.updateAddressesAddIcon(this.numberOfAddresses);
 
             // when  contact-form is initalized
             this.sandbox.on('sulu.contact-form.initialized', function() {
                 // set form data
-                var formObject = this.sandbox.form.create(this.form);
+                var formObject = this.sandbox.form.create(constants.formSelector);
                 formObject.initialized.then(function() {
                     this.formInitializedHandler(data);
                 }.bind(this));
@@ -306,13 +292,13 @@ define([
 
         setFormData: function(data) {
             // add collection filters to form
-            this.sandbox.emit('sulu.contact-form.add-collectionfilters', this.form);
+            this.sandbox.emit('sulu.contact-form.add-collectionfilters', constants.formSelector);
 
             this.numberOfBankAccounts = !!data.bankAccounts ? data.bankAccounts.length : 0;
             this.updateBankAccountAddIcon(this.numberOfBankAccounts);
 
-            this.sandbox.form.setData(this.form, data).then(function() {
-                this.sandbox.start(this.formContactFields);
+            this.sandbox.form.setData(constants.formSelector, data).then(function() {
+                this.sandbox.start(constants.formContactFields);
                 this.sandbox.emit('sulu.contact-form.add-required', ['email']);
                 this.sandbox.emit('sulu.contact-form.content-set');
                 this.dfdFormIsSet.resolve();
@@ -394,8 +380,8 @@ define([
         },
 
         save: function() {
-            if (this.sandbox.form.validate(this.form)) {
-                var data = this.sandbox.util.extend(false, {}, this.data, this.sandbox.form.getData(this.form));
+            if (this.sandbox.form.validate(constants.formSelector)) {
+                var data = this.sandbox.util.extend(false, {}, this.data, this.sandbox.form.getData(constants.formSelector));
                 if (!data.id) {
                     delete data.id;
                 }
@@ -407,9 +393,9 @@ define([
                 this.sandbox.emit('sulu.tab.saving');
                 AccountManager.save(data).then(function(savedData) {
                     this.data = savedData;
-                    this.initContactData();
-                    this.setFormData(this.data);
-                    this.sandbox.emit('sulu.tab.saved', savedData);
+                    var formData = this.initContactData();
+                    this.setFormData(formData);
+                    this.sandbox.emit('sulu.tab.saved', savedData, true);
                 }.bind(this));
             }
         },