+++ CHANGELOG.md
@@ -2,9 +2,10 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * FEATURE     #1478 [ContactBundle]   Rest-Api for media lists for contacts and accounts
     * FEATURE     #1421 [AdminBundle]     New button-api for header and tabs
     * FEATURE     #1417 [AdminBundle]     Style upgrade for header and tabs
-    * FEATURE     #1462 [ContactBundle]   Rest-Api support for contact-avatars
+    * FEATURE     #1472 [ContactBundle]   Rest-Api support for contact-avatars
     * BUGFIX      #1462 [Rest]            Fixed type of returned value for the Doctrine list builder count method
     * BUGFIX      #1437 [SnippetBundle]   Fixed copy-locale overlay bug
     * FEATURE     #1424 [All]             Implemented and integrated expressions for the listbuilder